refactor: rename to vscode
