+++ src/web/app/composables/index.ts
@@ -1,4 +1,2 @@
-export * from "./types";
 export * from "./utils";
-export * from "./StateController";
-export * from "./MockedStateController";
+export * from "./vscode";
